chore: split sidebar components and styles
