+++ app/components/layout/Layout.js
@@ -6,8 +6,8 @@ import { useMappedState, useDispatch } from 'redux-react-hook';
 import { withStyles } from '@material-ui/core/styles';
 import { MuiThemeProvider } from '@material-ui/core/styles';
 
-import TopBar from 'components/topbar/TopBarContainer';
-import Sidebar from 'components/sidebar/SidebarContainer';
+import TopBar from 'components/topbar/Container';
+import Sidebar from 'components/sidebar/Container';
 import Navigation from 'components/navigation/Navigation';
 import Packages from '../packages/packages/Packages';
 import Notifications from 'components/notifications/list/Container';
