+++ app/components/sidebar/tabs/actions/styles.js
