+++ app/components/sidebar/tabs/history/styles.js
