+++ app/components/sidebar/tabs/index.js
@@ -1,5 +1,5 @@
-import PackagesTab from './Packages';
-import HistoryTab from './History';
-import ActionsTab from './Actions';
+import PackagesTab from './packages/Packages';
+import HistoryTab from './history/History';
+import ActionsTab from './actions/Actions';
 
 export { ActionsTab, PackagesTab, HistoryTab };
