+++ app/components/sidebar/tabs/packages/styles.js
