+++ app/components/sidebar/tabs/styles.js
